+++ framework/CHANGELOG.md
@@ -10,6 +10,7 @@ Yii Framework 2 Change Log
 - Bug #20256: Add support for dropping views in MSSQL server when running migrate/fresh (ambrozt)
 - Enh #20248: Add support for attaching behaviors in configurations with Closure (timkelty)
 - Enh #20268: Minor optimisation in `\yii\helpers\BaseArrayHelper::map` (chriscpty)
+- Enh #20268: Add `\yii\helpers\BaseStringHelper::contains()` as a polyfill for `str_contains` (chriscpty)
 
 2.0.51 July 18, 2024
 --------------------